+++ agent/engine/docker_image_manager_integ_test.go
@@ -81,24 +81,17 @@ func TestIntegImageCleanupHappyCase(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	imageState1 := imageManager.GetImageStateFromImageName(test1Image1Name)
-	if imageState1 == nil {
-		t.Fatalf("Could not find image state for %s", test1Image1Name)
-	} else {
-		t.Logf("Found image state for %s", test1Image1Name)
-	}
-	imageState2 := imageManager.GetImageStateFromImageName(test1Image2Name)
-	if imageState2 == nil {
-		t.Fatalf("Could not find image state for %s", test1Image2Name)
-	} else {
-		t.Logf("Found image state for %s", test1Image2Name)
-	}
-	imageState3 := imageManager.GetImageStateFromImageName(test1Image3Name)
-	if imageState3 == nil {
-		t.Fatalf("Could not find image state for %s", test1Image3Name)
-	} else {
-		t.Logf("Found image state for %s", test1Image3Name)
-	}
+	imageState1, ok := imageManager.GetImageStateFromImageName(test1Image1Name)
+	require.True(t, ok, "Could not find image state for %s", test1Image1Name)
+	t.Logf("Found image state for %s", test1Image1Name)
+
+	imageState2, ok := imageManager.GetImageStateFromImageName(test1Image2Name)
+	require.True(t, ok, "Could not find image state for %s", test1Image2Name)
+	t.Logf("Found image state for %s", test1Image2Name)
+
+	imageState3, ok := imageManager.GetImageStateFromImageName(test1Image3Name)
+	require.True(t, ok, "Could not find image state for %s", test1Image3Name)
+	t.Logf("Found image state for %s", test1Image3Name)
 
 	imageState1ImageID := imageState1.Image.ImageID
 	imageState2ImageID := imageState2.Image.ImageID
@@ -196,24 +189,17 @@ func TestIntegImageCleanupThreshold(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	imageState1 := imageManager.GetImageStateFromImageName(test2Image1Name)
-	if imageState1 == nil {
-		t.Fatalf("Could not find image state for %s", test2Image1Name)
-	} else {
-		t.Logf("Found image state for %s", test2Image1Name)
-	}
-	imageState2 := imageManager.GetImageStateFromImageName(test2Image2Name)
-	if imageState2 == nil {
-		t.Fatalf("Could not find image state for %s", test2Image2Name)
-	} else {
-		t.Logf("Found image state for %s", test2Image2Name)
-	}
-	imageState3 := imageManager.GetImageStateFromImageName(test2Image3Name)
-	if imageState3 == nil {
-		t.Fatalf("Could not find image state for %s", test2Image3Name)
-	} else {
-		t.Logf("Found image state for %s", test2Image3Name)
-	}
+	imageState1, ok := imageManager.GetImageStateFromImageName(test2Image1Name)
+	require.True(t, ok, "Could not find image state for %s", test2Image1Name)
+	t.Logf("Found image state for %s", test2Image1Name)
+
+	imageState2, ok := imageManager.GetImageStateFromImageName(test2Image2Name)
+	require.True(t, ok, "Could not find image state for %s", test2Image2Name)
+	t.Logf("Found image state for %s", test2Image2Name)
+
+	imageState3, ok := imageManager.GetImageStateFromImageName(test2Image3Name)
+	require.True(t, ok, "Could not find image state for %s", test2Image3Name)
+	t.Logf("Found image state for %s", test2Image3Name)
 
 	imageState1ImageID := imageState1.Image.ImageID
 	imageState2ImageID := imageState2.Image.ImageID
@@ -333,8 +319,8 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	require.NoError(t, err, "task1")
 
 	// Verify image state is updated correctly
-	imageState1 := imageManager.GetImageStateFromImageName(identicalImageName)
-	require.NotNil(t, imageState1, "Could not find image state for %s", identicalImageName)
+	imageState1, ok := imageManager.GetImageStateFromImageName(identicalImageName)
+	require.True(t, ok, "Could not find image state for %s", identicalImageName)
 	t.Logf("Found image state for %s", identicalImageName)
 	imageID1 := imageState1.Image.ImageID
 
@@ -348,8 +334,8 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	require.NoError(t, err, "task2")
 
 	// Verify image state is updated correctly
-	imageState2 := imageManager.GetImageStateFromImageName(identicalImageName)
-	require.NotNil(t, imageState2, "Could not find image state for %s", identicalImageName)
+	imageState2, ok := imageManager.GetImageStateFromImageName(identicalImageName)
+	require.True(t, ok, "Could not find image state for %s", identicalImageName)
 	t.Logf("Found image state for %s", identicalImageName)
 	imageID2 := imageState2.Image.ImageID
 	require.NotEqual(t, imageID2, imageID1, "The image id in task 2 should be different from image in task 1")
@@ -364,8 +350,8 @@ func TestImageWithSameNameAndDifferentID(t *testing.T) {
 	require.NoError(t, err, "task3")
 
 	// Verify image state is updated correctly
-	imageState3 := imageManager.GetImageStateFromImageName(identicalImageName)
-	require.NotNil(t, imageState3, "Could not find image state for %s", identicalImageName)
+	imageState3, ok := imageManager.GetImageStateFromImageName(identicalImageName)
+	require.True(t, ok, "Could not find image state for %s", identicalImageName)
 	t.Logf("Found image state for %s", identicalImageName)
 	imageID3 := imageState3.Image.ImageID
 	require.NotEqual(t, imageID3, imageID1, "The image id in task3 should be different from image in task1")
@@ -463,8 +449,8 @@ func TestImageWithSameIDAndDifferentNames(t *testing.T) {
 	err = verifyTaskIsRunning(stateChangeEvents, task1)
 	require.NoError(t, err)
 
-	imageState1 := imageManager.GetImageStateFromImageName(task1.Containers[0].Image)
-	require.NotNil(t, imageState1, "Could not find image state for %s", task1.Containers[0].Image)
+	imageState1, ok := imageManager.GetImageStateFromImageName(task1.Containers[0].Image)
+	require.True(t, ok, "Could not find image state for %s", task1.Containers[0].Image)
 	t.Logf("Found image state for %s", task1.Containers[0].Image)
 	imageID1 := imageState1.Image.ImageID
 
@@ -481,8 +467,8 @@ func TestImageWithSameIDAndDifferentNames(t *testing.T) {
 	err = verifyTaskIsRunning(stateChangeEvents, task2)
 	require.NoError(t, err)
 
-	imageState2 := imageManager.GetImageStateFromImageName(task2.Containers[0].Image)
-	require.NotNil(t, imageState2, "Could not find image state for %s", task2.Containers[0].Image)
+	imageState2, ok := imageManager.GetImageStateFromImageName(task2.Containers[0].Image)
+	require.True(t, ok, "Could not find image state for %s", task2.Containers[0].Image)
 	t.Logf("Found image state for %s", task2.Containers[0].Image)
 	imageID2 := imageState2.Image.ImageID
 	require.Equal(t, imageID2, imageID1, "The image id in task2 should be same as in task1")
@@ -500,8 +486,8 @@ func TestImageWithSameIDAndDifferentNames(t *testing.T) {
 	err = verifyTaskIsRunning(stateChangeEvents, task3)
 	assert.NoError(t, err)
 
-	imageState3 := imageManager.GetImageStateFromImageName(task3.Containers[0].Image)
-	require.NotNil(t, imageState3, "Could not find image state for %s", task3.Containers[0].Image)
+	imageState3, ok := imageManager.GetImageStateFromImageName(task3.Containers[0].Image)
+	require.True(t, ok, "Could not find image state for %s", task3.Containers[0].Image)
 	t.Logf("Found image state for %s", task3.Containers[0].Image)
 	imageID3 := imageState3.Image.ImageID
 	require.Equal(t, imageID3, imageID1, "The image id in task3 should be the same as in task1")